+++ src/main/java/model/member/SellerDO.java
@@ -8,7 +8,7 @@ public class SellerDO {
     private String nickname;
     private String passwd;
     private int point;
-    private int tel;
+    private long tel;
     private String regdate;
     private String sellerImg;
     private String address;
@@ -61,11 +61,11 @@ public class SellerDO {
         this.point = point;
     }
 
-    public int getTel() {
+    public long getTel() {
         return tel;
     }
 
-    public void setTel(int tel) {
+    public void setTel(long tel) {
         this.tel = tel;
     }
